Keeping echidna-shim as a separate file.
The full echidna-shim binary was previously compiled in to echidna-lib as a CONST [u8]. This meant no external files were required, which is particularly helpful because then echidna-cli can be a standalone binary. Cargo was never really happy about this dependency, and it seems like it would be a problem for distributing the crates, so now echidna-shim is kept separate. This doesn't affect Echidna.app, which can keep a copy of echidna-shim in Resources/, but echidna-cli will need to be passed a path to echidna-shim (which the user will need to stash somewhere).