App: improving filename remembering semantics.
Previously, any filename entered in to the Save As dialog would be kept for the remainder of the session for any additional Save As clicks. This leads to the scenario where after using the default filenam, entering a second command and clicking Save As uses the default name from the previous command. Now, if the name used is the default, it is not saved, and any changes to the command will be reflected in the new default.